requirements.txt
# GPU-optimized requirements for influencer content chatbot creator

# Core PyTorch with CUDA support
torch>=2.1.0+cu121
torchvision>=0.16.0+cu121
torchaudio>=2.1.0+cu121

# Transformers and related libraries
transformers>=4.36.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Diffusion models
diffusers>=0.24.0

# Computer vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Utilities
requests>=2.31.0
numpy>=1.24.0
huggingface-hub>=0.19.0

# Performance monitoring
psutil>=5.9.0
nvidia-ml-py3>=7.352.0

# Optional optimizations
xformers>=0.0.22  # Memory efficient attention
safetensors>=0.4.0  # Faster model loading

# Development and logging
logging>=0.4.9.6
